chore: move OlapTable to its own file
